requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
Django==5.2.6
django-cloudinary-storage==0.3.0
django-debug-toolbar==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
